abstract out h2oModelBuilderForm
